# scholar/openapi.yaml
openapi: 3.1.0
info:
  title: Open Scholar API
  description: |
    Open Scholar is a decentralized platform for academic publishing and knowledge sharing,
    built on top of GrabNet - a peer-to-peer content-addressed network.
    
    ## Authentication
    Most endpoints require authentication via Bearer token. Obtain a token by logging in
    via `/api/auth/login`. Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <your-token>
    ```
    
    ## Rate Limits
    - General: 100 requests/second
    - Login: 10 attempts/minute per IP
    - Registration: 5 attempts/hour per IP
    - Uploads: 20/minute per user
  version: 0.1.0
  contact:
    name: Rooted Revival
    email: michaelwillisbotany@proton.me
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8889/api
    description: Local development
  - url: https://scholar.rootedrevival.us/api
    description: Production
...
paths:
...
  /files:
    post:
      tags: [Files]
      summary: Upload file
      description: |
        Upload a file with metadata. File is automatically pinned to GrabNet.
        Supports any file type.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload
                metadata:
                  type: string
                  description: JSON string with title, description, tags, is_public
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '400':
          description: Invalid file or metadata
        '401':
          description: Not authenticated
        '413':
          description: File too large
  /files/{uuid}:
    get:
      tags: [Files]
      summary: Get file metadata
      description: Get metadata for a specific file
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: File metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '404':
          description: File not found
    put:
      tags: [Files]
      summary: Update file metadata
      description: Update title, description, tags, or visibility
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFileRequest'
      responses:
        '200':
          description: File updated
        '401':
          description: Not authenticated
        '403':
          description: Not file owner
        '404':
          description: File not found
    delete:
      tags: [Files]
      summary: Delete file
      description: Delete a file (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: File deleted
        '401':
          description: Not authenticated
        '403':
          description: Not file owner
        '404':
          description: File not found
...
  /admin/stats:
    get:
      tags: [Admin]
      summary: Dashboard statistics
      description: Get overview statistics for admin dashboard
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStats'
        '401':
          description: Not authenticated
        '403':
          description: Not an admin
...
  /admin/files:
    get:
      tags: [Admin]
      summary: List all files
      description: Get paginated list of all files (admin view)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: File list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileList'
        '403':
          description: Not an admin
  /admin/files/{file_uuid}:
    delete:
      tags: [Admin]
      summary: Admin delete file
      description: Delete any file (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: file_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: File deleted
        '403':
          description: Not an admin
        '404':
          description: File not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token from login
  parameters:
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
      description: Number of results to return
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
      description: Number of results to skip
  schemas:
...
    File:
      type: object
      properties:
        id:
          type: integer
        uuid:
          type: string
          format: uuid
        user_id:
          type: integer
        username:
          type: string
        filename:
          type: string
        original_filename:
          type: string
        content_type:
          type: string
        size:
          type: integer
        hash:
          type: string
        grabnet_cid:
          type: string
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        is_public:
          type: boolean
        view_count:
          type: integer
        download_count:
          type: integer
        review_count:
          type: integer
        average_rating:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FileList:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
        total:
          type: integer
    UpdateFileRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        is_public:
          type: boolean
...
    AdminStats:
      type: object
      properties:
        total_users:
          type: integer
        total_files:
          type: integer
        total_reviews:
          type: integer
        total_storage:
          type: integer
        new_users_today:
          type: integer
        new_files_today:
          type: integer
        active_users_week:
          type: integer